[be] fix: FeedbackResponseDto 수정
