+++ server/src/main/java/com/CreatorConnect/server/feedbackcategory/dto/FeedbackCategoryDto.java
@@ -10,13 +10,11 @@ import javax.validation.constraints.NotBlank;
 
 public class FeedbackCategoryDto {
     @AllArgsConstructor
-    @NoArgsConstructor
     @Getter
-    @Setter
     public static class Post{
         @Valid
         @NotBlank(message = "카테고리를 입력하세요.")
-        private String category;
+        private String feedbackCategoryName;
 
     }
     @AllArgsConstructor
